requirements.txt
blinker==1.6.3
click==8.1.7
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyserial==3.5
werkzeug==3.0.1
zipp==3.17.0
